NeuralTrace/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.25.0
pydantic>=2.0.0
pyverilog>=1.3.0
openai>=1.0.0
google-genai>=0.3.0
python-dotenv>=1.0.0
networkx>=3.0.0
gunicorn>=21.0.0
